trace-cmd report: Make filter arguments match their files
Currently the filter argument "-F" only affects the first file passed in,
and is ignored for the rest of the files passed in via '-i'. Also, buffers
ignore them as well.

Have the -F affect the last '-i file', with the exception that the -F
filter can be called before the first file. That is, the first -i file is
affected by any -F before it and directly after it but before another -i
file is added.

Also, have buffers within a file be affected by the filters given to the
file.

Link: https://lore.kernel.org/linux-trace-devel/20220805154040.2014381-9-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>