tracecmd: Add NO_LIBZSTD option to disable libzstd
Other Linux kernel tools like perf already include options to disable
libzstd manually. Add this option also to trace-cmd.

The OpenWrt SDK fails to build trace-cmd due to the autodetection of
libzstd. The package is present in some feed repositories but should
not be used by trace-cmd. The compilation will fail with:

  Package trace-cmd is missing dependencies for the following libraries:
  libzstd.so.1

Link: https://github.com/rostedt/trace-cmd/pull/19

Signed-off-by: Nick Hainke <vincent@systemli.org>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>