trace-cmd: libtracecmd: Fixing linking to C++ code
Linking in C++ compilers (including g++) causes references to be created
with their arguments. Due to this, trace library headers included into
C++ code base will cause their objects to built with symbols with
arguments. Apparently this is to support operator overloading in C++.

This causes linker errors. For example, here's what I get when I try to
link libtracecmd with a main.o built from a C++ main.cc source file.

 main.cc:(.text+0x90):
	undefined reference to `tracecmd_get_tep(tracecmd_input*)'
	undefined reference to `tracecmd_free_record(tep_record*)'
	undefined reference to `tracecmd_read_data(tracecmd_input*, int)'

The standard fix for this is to wrap the C project's header in
extern "C".

With this patch, I am able to link libtracecmd into a C++
code base.

Link: https://lore.kernel.org/linux-trace-devel/20220815031342.775201-1-joel@joelfernandes.org

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>