trace-cmd report: Add callback for kvm plugin to show guest functions
Starting with libtraceevent 1.6.2, the kvm plugin will call a weak
function that can be overridden by the application (in this case
trace-cmd) called tep_plugin_kvm_get_func() that passes in the event, the
record, and an address to a value containing the address of the function
being requested. If the application has the guest mapping of addresses to
function names (usually taken from kallsyms), then it can define this
function to return the name of the function at the given address of the
guest.

Optionally, it can update the value holding the address, to the start of
the function such that the tep kvm plugin can print not only the function
name, but also the address offset of the original location and the
function itself.

Link: https://lore.kernel.org/linux-trace-devel/20220921202642.44fec719@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>