trace-cmd record: Use result of fcntl(GETPIPE_SZ)
For some reason, I thought the arg to fcntl(fd, F_GETPIPE_SZ, &arg) would
give the pipe size. But the argument is completely ignored (I remember
testing this, and it worked :-/)

The return of the function is the pipe size. Use that if pipe_size is
untouched (which it ought to be) as the pipe size.

Link: https://lore.kernel.org/linux-trace-devel/20220928095753.54b6621c@gandalf.local.home

Fixes: a64fab23cec2 ("trace-cmd recorder: Use full pipe size in splice_data()")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>