trace-cmd library: Do not call callbacks with NULL record
The callbacks path can possibly be called when the next record is NULL. Do
not call the callbacks in this case, as they are not expecting a NULL
record, nor should they have to.

Move the code that calls the callbacks in tracecmd_iterate_events() and
tracecmd_iterate_events_multi() into a single helper function to alleviate
fixing one and not the other.

Link: https://lore.kernel.org/linux-trace-devel/20221007143859.6895d756@gandalf.local.home

Reported-by: Ian Rogers <irogers@google.com>
Fixes: 3cd1b55f3 ("trace-cmd library: Add tracecmd_follow_event()")
Fixes: 82ed4a937 ("trace-cmd library: Add filtering logic for iterating events")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>