trace-cmd library: Update record after callback in iterators
The function graph plugin will look at the next event and may even
increment it if it is a leaf function. The current iterators update the
next record before the callback. This leaves them vulnerable to
corruption if the callback updates to the next record like function
graph plugin does.

Peek at the next record after the callback to ensure that the next
record is the one that will be used for the next iterator.

Link: https://lore.kernel.org/linux-trace-devel/20221110003325.007663c7@rorschach.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>