trace-cmd: Fix reading stats of read max test
The read max test checks the stats by running trace-cmd report --stat on
the file. The problem is that it left out the output file that was
created, and if the default file trace.dat does not exist, it fails to
read it. Unfortunately, if it does exist, it may give a false pass or
fail.

Link: https://lore.kernel.org/linux-trace-devel/20221210205328.1f1f03ce@gandalf.local.home

Fixes: 1f75010a2 ("trace-cmd: Add test for max size option of record")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>