trace-cmd library: Fix read_data() with error from tracefs_cpu_read()
If tracefs_cpu_read() has an error, it will return a negative value. But the
value is then used as an index. An error check is required to make sure that
if an error is returned from tracefs_cpu_read() the function exits with that
error and does not continue further.

Link: https://lore.kernel.org/linux-trace-devel/20230106183930.12565-6-rostedt@goodmis.org

Fixes: 2610bdcde ("trace-cmd library: Use tracefs_cpu for recorder helpers")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>