trace-cmd report: Ensure filter is applied to single input file
Since 955d05fc7aee ("trace-cmd report: Make filter arguments match their
files"), the -F filtering is silently ignored when a trace file is
provided with -i and the filter comes after -i .  The reason is that the
filter is now associated with input_files and not saved to the global
list only in this case, but process_filters still only checks the global
list when handles->input_file is not set.

Avoid this by checking last_input_file first, which always contains a
pointer to the correct filter in this case.

Link: https://lore.kernel.org/linux-trace-devel/20230330185211.5604-2-krisman@suse.de

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=217038
Fixes: 955d05fc7aee ("trace-cmd report: Make filter arguments match their files")
Signed-off-by: Gabriel krisman Bertazi <krisman@suse.de>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>