libtracecmd: Fix tracecmd_compress_copy_from() write size return
The amount written to the file returned by tracecmd_compress_copy_from()
is off by 4 bytes. This is because it did not account for the writing of
the number of chunks saved.

Instead of calculating the amount written to the file, as lseek() is used to
find the chunk value and also put it back to the end of what was written,
use the results of lseek() to figure out the amount written.

It appears that the decompression of the sections did not use the size of
the buffer section, but instead used the size inside each chunk. That is,
it read the chunks within the buffer section and based what it would read
from the file by the size of each chunk, and did not use the value defined
in the section header.

Link: https://lore.kernel.org/linux-trace-devel/20230531055401.1f04a9fd@rorschach.local.home

Cc: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>
Reported-by: Douglas Raillard <douglas.raillard@arm.com>
Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=217367
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>