tracecmd library: Unlock records in tracecmd_iterate_events()
The tracecmd_iterate_events() and tracecmd_iterate_events_multi() uses
tracecmd_peek_data() to look at the next record. But when this is done,
the record is "cached" and "locked" in the handle. Which means they can
not be freed.

At the end of the iterators, make sure to read the data to unlock them,
and then free them.

Link: https://lore.kernel.org/linux-trace-devel/20230606145231.77820ea6@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>