trace-cmd library: Have callbacks exit out of the iterator
In the man page for tracecmd_iterate_events() it states that if the
callback() returns anything other than zero, it will exit the iterator.
But this is not true. It only exits if the callback returns less than
zero. This is a bug as the code should do what the man page states. Have
it exit the loop if the callback returns non zero.

Link: https://lore.kernel.org/linux-trace-devel/20230615131135.123e38af@gandalf.local.home

Fixes: 2cb6cc2f4e556 ("tracecmd library: Add tracecmd_iterate_events()")
Fixes: b37903ae5fb53 ("tracecmd library: Add tracecmd_iterate_events_multi()")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>