trace-cmd split: Fix creation of temp files to include output name
The creation of the temp files for creating the split did the following:

	output = strdup(output_file);
	dir = dirname(output);
	base = basename(output);

The problem with the above is that dirname() modifies "output" to add a
'\0' after the directory name. That is, if we had output_file = '/tmp/test';
Then the following will happen:

	output = copy of "/tmp/test"
	dir = dirname(output);

This will set dir to "/tmp" and insert a '\0' into the second '/'. This
also has the side effect of changing output to "/tmp" as well.

Then to get the base name:

	base = basename(output);

base will end up as "tmp" and not as "test". The temp file will then be:

		ret = asprintf(&file, "%s/.tmp.%s.%d", dir, base, cpu);

	"/tmp/.tmp.tmp.1"

This is incorrect, as it should have been "/tmp/.tmp.test.1".

This can easily be fixed by swapping the assignment of dir and base:

	base = basename(output);
	dir = dirname(output);

As then, after the assignment of base, base would be "test" and output
would still be "/tmp/test", allowing for dir to get "/tmp".

Link: https://lore.kernel.org/linux-trace-users/VE1PR09MB3535122A0DA44242886307B9F233A@VE1PR09MB3535.eurprd09.prod.outlook.com/
Link: https://lore.kernel.org/linux-trace-devel/20230710125031.0e140727@gandalf.local.home

Reported-by: Sharon Gabay <Sharon.Gabay@mobileye.com>
Fixes: d9b58d42c43c2 ("trace-cmd: Fix temp files in trace-cmd split to include directories")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>