trace-cmd record: Use SIGTERM as well for --daemonize
Currently --daemonize expects a SIGINT to stop the daemon and build the
trace.dat file. But if the user just does a:

  kill <pid>

it will send a SIGTERM. This will actually kill the process and it will
not create the trace.dat file, leaving a bunch of trace.dat.cpuX files
just lying around.

Update it to also do the same with SIGTERM.

Link: https://lore.kernel.org/linux-trace-devel/20231017233503.2205514-2-rostedt@goodmis.org

Cc: Avidan Borisov <avidanborisov@gmail.com>
Cc: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>