trace-cmd record: Use copy of PATH for strtok_r() operations
As strtok_r() modifies the string it is parsing, using the environment variable
PATH to find the paths for execution causes it to be truncated when reused by
exec. Instead, make a copy of the PATH environment variable to use to parse the
paths.

I had this fixed in my repo for some time and never pushed it out, but it was
eventually reported by others.

Link: https://lore.kernel.org/all/20231128192435.36507-1-void@manifault.com/
Link: https://lore.kernel.org/linux-trace-devel/20231128153024.6e0d40e1@gandalf.local.home

Cc: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Reported-by: David Vernet <void@manifault.com>
Reviewed-by: David Vernet <void@manifault.com>
Fixes: edf9424029cc ("trace-cmd: Open code execvp routine to avoid multiple execve syscalls")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>