trace-cmd library: Add back setting errno to zero in tracecmd_stack_tracer_status()
A clean up was made to remove all the settings of errno to zero in the
library, but it was a bit too aggressive in doing so. It removed the errno
setting to zero in tracecmd_stack_tracer_status() just before it called
strtol() which it uses to test for the success of that function. By not
clearing errno, it was causing trace-cmd stat to report:

 "Error reading stack tracer status"

Link: https://lore.kernel.org/linux-trace-devel/20231229031138.68313-3-rostedt@goodmis.org

Fixes: dcd5ee3142 ("libtracecmd: Do not set errno to zero")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>