trace-cmd report: Add interrupt info in output
In the Linux kernel version 3.3 the default output showed the irq information
that is typically only for the latency format. That is, it shows:

               <idle>-0     [000] d..2    49.309305: cpuidle_get_driver <-cpuidle_idle_call
               <idle>-0     [000] d..2    49.309307: mwait_idle <-cpu_idle
               <idle>-0     [000] d..2    49.309309: need_resched <-mwait_idle

Instead of:

               <idle>-0     [000]     49.309305: cpuidle_get_driver <-cpuidle_idle_call
               <idle>-0     [000]     49.309307: mwait_idle <-cpu_idle
               <idle>-0     [000]     49.309309: need_resched <-mwait_idle

After 12 years, it's about time that the default trace-cmd report output did
the same!

This does change the output format and possibly can break anything that
tries to parse it.

Link: https://lore.kernel.org/linux-trace-devel/20240108191759.5fc4ecda@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>