libtracecmd: Fix free_zpage() offset
The offset exposed to the callers of libtracecmd is the uncompressed file
offset. The read_zpage() subtracts the fake uncompressed offset to get to
the compressed data offset.

The free_zpage() did not subtract the uncompressed offset so it was not
finding the cached compressed data it was looking for to free, and left the
page not freed.

Subtract the uncompressed file offset from the passed in offset to find the
cached node to free.

Link: https://lore.kernel.org/linux-trace-devel/20240111171207.54d7cdbf@gandalf.local.home

Fixes: add83e0c8b511 ("trace-cmd library: Fix tracecmd_read_at()")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>