trace-cmd record: Fix top_instance.output_file being NULL
If the -o comes after a -B, the output_file name will be associated to the
trace instance of that -B, and the top_instance.output_file will be NULL.
If recording is started in a location that is read-only, the temp file
created for the top instance is going to try to be created in that read-only
file system.

 /sys/kernel/tracing# trace-cmd record -o /tmp/test.dat -B test -e sched
 Hit Ctrl^C to stop recording
 ^Ctrace-cmd: Permission denied
   could not create file (null).cpu0

Make sure all instances, including the top_instance has their output_file
correct.

Also, because the output file could have been created under another
instance, check if that instance already has the output file before
assigning over it.

Link: https://lore.kernel.org/linux-trace-devel/20240119162743.1a107fa9@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>