trace-cmd report: Allow for multiple files without having -i
If trace-cmd report is called with a single file that is not specified by
the -i option, it will report it. But if it is called with multiple files,
it will only report the first one. Allow all the trace.dat files at the
end of options to be considered files to read.

That is, currently if you have:

  trace-cmd record -e sched sleep 100
  trace-cmd split -r -m 100 -o trace-split.dat
  trace-cmd report trace-split.dat.*

It will only show the report of trace-split.dat.0001 and not the rest. To
get the rest, each one would need to be included with -i. Instead, allow
them all to be reported with the above command.

Link: https://lore.kernel.org/linux-trace-devel/20240124132231.638e148a@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>